# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SourceIt API",
    "version": "0.1.0",
    "description": "Generated from packages/shared/src/zod — not hand-written. See docs/PROJECT_STATE.md for what is and isn't implemented yet."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ]
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "Evidence": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "b3f1c9a0-1e2d-4a3b-9c8f-6d5e4f3a2b1c"
          },
          "articleVersionId": {
            "type": "string",
            "format": "uuid",
            "example": "b3f1c9a0-1e2d-4a3b-9c8f-6d5e4f3a2b1c"
          },
          "fileType": {
            "type": "string",
            "enum": [
              "image",
              "video",
              "document"
            ]
          },
          "tag": {
            "type": "string",
            "enum": [
              "cover_image",
              "media",
              "evidence",
              "source"
            ]
          },
          "filename": {
            "type": "string"
          },
          "caption": {
            "type": [
              "string",
              "null"
            ]
          },
          "contentHash": {
            "type": "string"
          },
          "sourceUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "isArchivedSnapshot": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-15T14:30:00Z"
          }
        },
        "required": [
          "id",
          "articleVersionId",
          "fileType",
          "tag",
          "filename",
          "caption",
          "contentHash",
          "sourceUrl",
          "isArchivedSnapshot",
          "createdAt"
        ]
      },
...
      "UploadEvidenceRequest": {
        "type": "object",
        "properties": {
          "fileType": {
            "type": "string",
            "enum": [
              "image",
              "video",
              "document"
            ]
          },
          "tag": {
            "type": "string",
            "enum": [
              "cover_image",
              "media",
              "evidence",
              "source"
            ]
          },
          "filename": {
            "type": "string",
            "minLength": 1
          },
          "caption": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "fileType",
          "tag",
          "filename"
        ]
      },
      "UploadEvidenceMultipart": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UploadEvidenceRequest"
          },
          {
            "type": "object",
            "properties": {
              "file": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/versions/{versionId}/evidence": {
      "get": {
        "tags": [
          "Evidence"
        ],
        "summary": "Evidence attached to this version (EvidenceSection.tsx)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3f1c9a0-1e2d-4a3b-9c8f-6d5e4f3a2b1c"
            },
            "required": true,
            "name": "versionId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Evidence"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No such published version (unknown, or still a draft)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Evidence"
        ],
        "summary": "Attach evidence to a draft version (MediaEvidenceUpload.tsx)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3f1c9a0-1e2d-4a3b-9c8f-6d5e4f3a2b1c"
            },
            "required": true,
            "name": "versionId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadEvidenceMultipart"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            }
          },
          "400": {
            "description": "Missing file part for a non-source tag, or malformed fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such version, or the caller may not write its draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version is no longer a draft — evidence binds while drafting and is then append-only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/versions/{versionId}/evidence/{evidenceId}/file": {
      "get": {
        "tags": [
          "Evidence"
        ],
        "summary": "Redirect to a short-lived signed URL for this evidence file (EvidenceSection.tsx 'View File'). Public, like the listing; the bucket itself is private and a fresh link is minted per call.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3f1c9a0-1e2d-4a3b-9c8f-6d5e4f3a2b1c"
            },
            "required": true,
            "name": "versionId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3f1c9a0-1e2d-4a3b-9c8f-6d5e4f3a2b1c"
            },
            "required": true,
            "name": "evidenceId",
            "in": "path"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to a signed URL, valid briefly"
          },
          "404": {
            "description": "No such published version, or no such evidence on it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}